rearrange(server): create auth controllers for auth routes
